refactor(ui): consolidate eventQueue tests, delete refactor_test.go
Move 4× TestEventQueue_EnqueueEvent_* from refactor_test.go to
event_queue_test.go, TestUIBridge_HandleEvent_SafetyWrapper to
bridge_test.go, and the misplaced TestUIBridge_HandleEvent_
CallerContextCancelled from event_queue_test.go to bridge_test.go.

Delete refactor_test.go (empty after moves). Add doc.go with
test file naming conventions.

All tests now live in exactly two files:
  - event_queue_test.go : unit tests (TestEventQueue_*)
  - bridge_test.go      : integration tests (TestUIBridge_*)

Closes #234
